% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_airtable.R
\name{query_airtable}
\alias{query_airtable}
\title{Queries Airtable}
\usage{
query_airtable(
 
    sheet_url = "https://docs.google.com/spreadsheets/d/1dCOcNwOYZDQXeyAHnc5ypDyaKmwm0OOKWq5nFAWR924/edit#gid=417127965"
)
}
\arguments{
\item{sheet_url}{URL of Airtable Google sheet}
}
\value{
Airtable tibble
}
\description{
This function queries the MD+ Airtable Directory.
}
\examples{
query_airtable()
}
